+++ mex/editor/edit/state.py
@@ -7,6 +7,7 @@ from requests import HTTPError
 
 from mex.common.backend_api.connector import BackendApiConnector
 from mex.common.logging import logger
+from mex.common.types import PublishingTarget
 from mex.editor.label_var import label_var
 from mex.editor.models import SearchResult
 from mex.editor.rules.state import RuleState
@@ -49,6 +50,16 @@ class EditState(RuleState):
                     await resolve_editor_value(preview)
 
         return results
+
+    @rx.event
+    def toggle_publish_target(self, publish_target_id: str) -> None:
+        """Toggle the given publish target by id."""
+        if self.workflow_rule:
+            target = PublishingTarget(publish_target_id)
+            if publish_target_id in self.workflow_rule.forbiddenPublishingTarget:
+                self.workflow_rule.forbiddenPublishingTarget.remove(target)
+            else:
+                self.workflow_rule.forbiddenPublishingTarget.append(target)
 
     @rx.event
     def delete_reset(self) -> Generator[EventSpec | None]:
@@ -94,37 +105,9 @@ class EditState(RuleState):
             if event := self.push_url_params(params):
                 yield event
 
-    @rx.var(cache=False)
-    def any_primary_source_or_editor_value_enabled(self) -> bool:
-        """Determine if any primary source or editor value is enabled.
-
-        Returns:
-            Return true if any primary source or editor value is enabled,
-            otherwise false.
-        """
-        return any(
-            ps.enabled or any(value.enabled for value in ps.editor_values)
-            for field in self.fields
-            for ps in field.primary_sources
-        )
-
-    @rx.event
-    def toggle_all_primary_source_and_editor_values(
-        self,
-    ) -> Generator[EventSpec]:
-        """Toggle all primary source and editor values."""
-        any_enabled = self.any_primary_source_or_editor_value_enabled
-        new_state = not any_enabled
-        for field in self.fields:
-            for ps in field.primary_sources:
-                ps.enabled = new_state
-                for value in ps.editor_values:
-                    value.enabled = new_state
-        yield RuleState.update_local_state  # type: ignore[misc]
-
-    @label_var(label_id="edit.toggle_all")
-    def label_toggle_all(self) -> None:
-        """Label for toggle_all."""
+    @label_var(label_id="edit.publish_targets")
+    def label_publish_targets(self) -> None:
+        """Label for publish_targets."""
 
     @label_var(label_id="edit.discard_changes.button")
     def label_discard_changes_button(self) -> None: